--- src/utils/apiClient.tsx
@@ -1,12 +1,0 @@
-import axios from 'axios';
-
-const BASE_URL = process.env.REACT_APP_BASE_URL || 'http://localhost:8080/api/v1';
-
-const apiClient = axios.create({
-  baseURL: BASE_URL,
-  timeout: 10000,
-});
-
-// TODO axios configuration e.g. interceptors
-
-export default apiClient;
